refactor: move getDateFormat helper to a shared location
Relocates the getDateFormat function from DateField component to a shared helpers directory. This change improves code organization and reusability, allowing multiple components to access this utility function from a centralized location. The logic for determining date format based on locale has been simplified to specifically check for 'en-US' locale, defaulting to 'dd/MM/yyyy' for all other locales.
